Update CDN deploy script to use new TBone build strategy from version 0.6.0 on

+++ buildcdn.sh
@@ -15,20 +15,28 @@ cd _tbone
 git fetch origin
 git checkout master
 git reset --hard origin/master
-for ref in `git ls-remote origin | grep ".*refs/[^p]" | sed "s/.*\///"`; do
+for ref in `git ls-remote origin | grep ".*refs/[^p]" | sed "s/.*\///" | sort -r`; do
     echo "Building tbone $ref"
     git checkout -q $ref
-    OPTIMIZATION_LEVEL=ADVANCED_OPTIMIZATIONS ./compile.py > ../_cdn/tbone-$ref.min.js
-    # compile with debug=true has the side effect of generating build/tbone.debug.js
-    TBONE_DEBUG=TRUE OPTIMIZATION_LEVEL=WHITESPACE_ONLY ./compile.py > /dev/null
-    cp build/tbone.debug.js ../_cdn/tbone-$ref.js
-    cp build/tbone.min.js.map ../_cdn/tbone-$ref.min.js.map
+    if [ -f "compile.py" ]; then
+        # Old build method, 0.5 and older
+        OPTIMIZATION_LEVEL=ADVANCED_OPTIMIZATIONS ./compile.py > ../_cdn/tbone-$ref.min.js
+        # compile with debug=true has the side effect of generating build/tbone.debug.js
+        TBONE_DEBUG=TRUE OPTIMIZATION_LEVEL=WHITESPACE_ONLY ./compile.py > /dev/null
+        cp build/tbone.debug.js ../_cdn/tbone-$ref.js
+        cp build/tbone.min.js.map ../_cdn/tbone-$ref.min.js.map
+    else
+        cp dist/tbone.js ../_cdn/tbone-$ref.js
+        cp dist/tbone.min.js ../_cdn/tbone-$ref.min.js
+        cp dist/tbone.min.js.map ../_cdn/tbone-$ref.min.js.map
+    fi
     sed -i s/tbone.min.js.map/tbone-$ref.min.js.map/ ../_cdn/tbone-$ref.min.js
     if [ -f "package.json" ]; then
         mkdir -p tbone/
-        cp build/tbone.debug.js tbone/tbone.js
+        cp ../_cdn/tbone-$ref.js tbone/tbone.js
         cp package.json LICENSE README.md tbone/
         sed -i "s/build\/tbone\.debug\.js/tbone\.js/" tbone/package.json
+        sed -i "s/dist\/tbone\.js/tbone\.js/" tbone/package.json
         tar czf ../_cdn/tbone-$ref.tgz tbone/
         rm -rf tbone/
     fi
